improved dragging on alignments
